Moved deactivate upwards
